+++ apps/docs/src/themes/theme.tsx
@@ -1,0 +1,59 @@
+import type { NextraThemeLayoutProps } from "nextra";
+import useBlogContext from "nextra-theme-blog";
+import React from "react";
+
+import {
+  ThemeMain,
+  ThemeHeader,
+  ThemeFooter,
+  PostsList,
+  NavList,
+  Nav,
+  NavBrand,
+} from "@fpkit/nextra";
+
+import { usePageOpts, useConfig } from "@fpkit/nextra/libs/hooks";
+import { FilterMdxPages } from "@fpkit/nextra/libs/libs";
+import { MDXProvider } from "nextra/mdx";
+import "@shawnsandy/first-paint/dist/css/libs/all.min.css";
+
+export default function Layout({
+  children,
+  pageOpts,
+  themeConfig,
+}: NextraThemeLayoutProps) {
+  const { pageMap } = pageOpts;
+  const { dirList, postList, pageList } = usePageOpts({
+    options: { pageOpts },
+  });
+
+  
+  const banner = themeConfig.banner;
+
+  const latest = FilterMdxPages(pageMap);
+
+  return (
+    <>
+      <Nav>
+        <NavBrand logo={themeConfig.brand.logo} url={themeConfig.brand.url} />
+      </Nav>
+
+      <ThemeHeader
+        title={banner.title}
+        link={banner.link}
+        linkLabel={banner.linkLabel}
+        description={banner.description}
+      />
+      <ThemeMain>
+        {pageOpts.route === "/" && latest.length > 0 ? (
+          <PostsList postList={latest} showDescription />
+        ) : (
+          <MDXProvider components={themeConfig.components}>
+            {children}
+          </MDXProvider>
+        )}
+      </ThemeMain>
+      <ThemeFooter />
+    </>
+  );
+}