+++ apps/next/src/app/about/page.tsx
@@ -1,0 +1,11 @@
+import { Main, Section } from "@fpkit/nextra";
+
+export default function Page() {
+  return (
+    <Main>
+      <Section>
+        <h2>Hello, Dashboard Page!</h2>
+      </Section>
+    </Main>
+  );
+}