+++ apps/nextra-blog/next.config.mjs
@@ -1,13 +1,14 @@
-import nextra from 'nextra'
+import nextra from "nextra";
 
 const withNextra = nextra({
-  theme: './src/themes/theme.tsx',
-  themeConfig: './theme.config.jsx',
+  theme: "./src/themes/theme.tsx",
+  themeConfig: "./theme.config.jsx",
   staticImage: true,
   defaultShowCopyCode: true,
-  readingTime: true
-})
+  flexsearch: true,
+  readingTime: true,
+});
 
 export default withNextra({
-  reactStrictMode: true
-})
+  reactStrictMode: true,
+});