+++ apps/nextra-blog/public/shiki.js
@@ -1,0 +1,10 @@
+shiki
+  .getHighlighter({
+    theme: "nord",
+  })
+  .then((highlighter) => {
+    const code = highlighter.codeToHtml(`console.log('shiki');`, {
+      lang: "js",
+    });
+    document.getElementById("output").innerHTML = code;
+  });