+++ packages/fpkit-nextjs/src/components/nav/nav.test.tsx
@@ -11,7 +11,6 @@ describe("App renders correctly", () => {
     const nav = screen.getByRole("navigation");
     expect(nav).toBeDefined();
     expect(await axe(nav)).toHaveNoViolations();
-    screen.debug();
     expect(nav).toMatchSnapshot();
   });
 });